requirements.txt
docker==7.1.0
flask==3.0.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1
bcrypt==4.2.0
flask-cors==5.0.0
PyJWT==2.9.0
gunicorn
